SPA_Backend/requirements.txt
Django>=4.2
psycopg2-binary
python-dotenv


djangorestframework
djangorestframework-simplejwt


django-cors-headers


django-filter


scikit-learn
numpy
pandas
joblib